Added password reset and cleaned up the code
